+++ R/esp_cache.R
@@ -1,9 +1,9 @@
-#' Set your `mapSpain` cache dir
+#' Set your **mapSpain** cache dir
 #'
 #' @family cache utilities
 #' @seealso [rappdirs::user_config_dir()]
 #'
-#' @return Invisible value
+#' @return An (invisible) character with the path to your `cache_dir`.
 #' @description
 #' This function will store your `cache_dir` path on your local machine and
 #' would load it for future sessions. Type `Sys.getenv("MAPSPAIN_CACHE_DIR")` to
@@ -14,15 +14,18 @@
 #'   * Run `Sys.setenv(MAPSPAIN_CACHE_DIR = "cache_dir")`. You would need to
 #'     run this command on each session (Similar to `install = FALSE`).
 #'   * Set `options(mapSpain_cache_dir = "cache_dir")`. Similar to the previous
-#'     option.
+#'     option. This is **not recommended any more**, and it is provided for
+#'     backwards compatibility purposes.
 #'   * Write this line on your .Renviron file:
-#'     `MAPSPAIN_CACHE_DIR = "cache_dir"` (same behavior than
+#'     `MAPSPAIN_CACHE_DIR = "value_for_cache_dir"` (same behavior than
 #'     `install = TRUE`). This would store your `cache_dir` permanently.
 #'
 #' @inheritParams esp_get_nuts
-#' @param cache_dir A path to a cache directory.
+#' @param cache_dir A path to a cache directory. On missing value the function
+#'   would store the cached files on a temporary dir (See [base::tempdir()]).
 #' @param install if `TRUE`, will install the key in your local machine for
-#'   use in future sessions.  Defaults to `FALSE.`
+#'   use in future sessions.  Defaults to `FALSE.` If `cache_dir` is `FALSE`
+#'   this parameter is set to `FALSE` automatically.
 #' @param overwrite If this is set to `TRUE`, it will overwrite an existing
 #'   `MAPSPAIN_CACHE_DIR` that you already have in local machine.
 #'
@@ -36,12 +39,30 @@
 #'
 #' Sys.getenv("MAPSPAIN_CACHE_DIR")
 #' @export
-esp_set_cache_dir <- function(cache_dir = rappdirs::user_cache_dir("mapSpain", "R"),
-                              overwrite = TRUE,
-                              install = TRUE,
+esp_set_cache_dir <- function(cache_dir,
+                              overwrite = FALSE,
+                              install = FALSE,
                               verbose = FALSE) {
 
   # nocov start
+
+  # Default if not provided
+  if (missing(cache_dir) || cache_dir == "") {
+    if (verbose) {
+      message(
+        "Using a temporary cache dir. ",
+        "Set 'cache_dir' to a value for store permanently"
+      )
+    }
+    # Create a folder on tempdir
+    cache_dir <- file.path(tempdir(), "mapSpain")
+    is_temp <- TRUE
+    install <- FALSE
+  } else {
+    is_temp <- FALSE
+  }
+
+
   # Validate
   stopifnot(
     is.character(cache_dir),
@@ -66,7 +87,7 @@ esp_set_cache_dir <- function(cache_dir = rappdirs::user_cache_dir("mapSpain", "
   # Install path on environ var.
 
   if (install) {
-    config_dir <- rappdirs::user_cache_dir("mapSpain", "R")
+    config_dir <- rappdirs::user_config_dir("mapSpain", "R")
     # Create cache dir if not presente
     if (!dir.exists(config_dir)) {
       dir.create(config_dir, recursive = TRUE)
@@ -79,25 +100,45 @@ esp_set_cache_dir <- function(cache_dir = rappdirs::user_cache_dir("mapSpain", "
       writeLines(cache_dir, con = mapspain_file)
     } else {
       stop(
-        "A cache_dir path already exists. You can overwrite it with the ",
-        "argument overwrite=TRUE",
+        "A cache_dir path already exists.\nYou can overwrite it with the ",
+        "argument overwrite = TRUE",
         call. = FALSE
       )
     }
   } else {
-    if (verbose) {
+    if (verbose && !is_temp) {
       message(
-        "To install your cache_dir path for use in future sessions, run this ",
+        "To install your cache_dir path for use in future sessions, \nrun this ",
         "function with `install = TRUE`."
       )
     }
   }
 
   Sys.setenv(MAPSPAIN_CACHE_DIR = cache_dir)
-  return(invisible())
+  return(invisible(cache_dir))
   # nocov end
 }
 
+esp_clear_cache <- function(config = TRUE,
+                            cached_data = TRUE,
+                            verbose = FALSE) {
+  config_dir <- rappdirs::user_config_dir("mapSpain", "R")
+  data_dir <- esp_hlp_detect_cache_dir()
+  if (config && dir.exists(config_dir)) {
+    unlink(config_dir, recursive = TRUE, force = TRUE)
+    if (verbose) message("mapSpain cache config deleted")
+  }
+
+  if (cached_data && dir.exists(data_dir)) {
+    unlink(data_dir, recursive = TRUE, force = TRUE, expand = TRUE)
+    if (verbose) message("mapSpain cached data deleted: ", data_dir)
+  }
+
+
+  Sys.setenv(MAPSPAIN_CACHE_DIR = "")
+  options(mapSpain_cache_dir = NULL)
+  return(invisible())
+}
 
 #' Detect cache dir for mapSpain
 #'
@@ -113,8 +154,8 @@ esp_hlp_detect_cache_dir <- function() {
   from_option <- getOption("mapSpain_cache_dir", NULL)
 
   if (!is.null(from_option) && (is.null(getvar) || getvar == "")) {
-    Sys.setenv(MAPSPAIN_CACHE_DIR = from_option)
-    return(from_option)
+    cache_dir <- esp_set_cache_dir(from_option, install = FALSE)
+    return(cache_dir)
   }
 
 
@@ -122,20 +163,22 @@ esp_hlp_detect_cache_dir <- function() {
 
   if (is.null(getvar) || is.na(getvar) || getvar == "") {
     # Not set - tries to retrieve from cache
-    cachedir <- rappdirs::user_cache_dir("mapSpain", "R")
-    mapspain_cache_path <- file.path(cachedir, "mapSpain_cache_dir")
+    cache_config <- file.path(
+      rappdirs::user_config_dir("mapSpain", "R"),
+      "mapSpain_cache_dir"
+    )
 
-    if (file.exists(mapspain_cache_path)) {
-      cached_path <- readLines(mapspain_cache_path)
+    if (file.exists(cache_config)) {
+      cached_path <- readLines(cache_config)
 
       # Case on empty cached path - would default
       if (is.null(cached_path) ||
         is.na(cached_path) || cached_path == "") {
-        esp_set_cache_dir(
-          install = TRUE, overwrite = TRUE,
+        cache_dir <- esp_set_cache_dir(
+          overwrite = TRUE,
           verbose = FALSE
         )
-        return(rappdirs::user_cache_dir("mapSpain", "R"))
+        return(cache_dir)
       }
 
       # 3. Return from cached path
@@ -144,11 +187,11 @@ esp_hlp_detect_cache_dir <- function() {
     } else {
       # 4. Default cache location
 
-      esp_set_cache_dir(
-        install = TRUE, overwrite = TRUE,
+      cache_dir <- esp_set_cache_dir(
+        overwrite = TRUE,
         verbose = FALSE
       )
-      return(rappdirs::user_cache_dir("mapSpain", "R"))
+      return(cache_dir)
     }
   } else {
     return(getvar)